+++ packages/python-packages/apiview-copilot/evals/run.py
@@ -215,7 +215,7 @@ def review_apiview(query: str, language: str):
         ApiViewReview,
     )
 
-    reviewer = ApiViewReview(target=query, language=language, base=None, debug_log=True)
+    reviewer = ApiViewReview(target=query, language=language, base=None)
     review = reviewer.run()
     reviewer.close()
     return {"actual": review.model_dump_json()}
